tools/pyproject.toml
[tool.poetry]
name = "translations"
version = "0.1.0"
description = ""
authors = ["Andrew Hankinson <andrew.hankinson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
ipython = "^7.18.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
